change PlayerData.alignment to long and add InverseLerp(long,long,long) to Utilities.cs

+++ Covens/Assets/Scripts/Game/Marker Managers/MarkerClasses.cs
@@ -216,7 +216,7 @@ public class PlayerData : WitchMarkerData
     public string account;
     public string physicalDominion;
     public ulong xp;
-    public int alignment;
+    public long alignment;
     public bool whiteMastery;
     public bool shadowMastery;
     public bool greyMastery;
